requirements.txt
PyYAML
pandas
scikit-learn
xgboost
catboost
shap
lime
matplotlib
deap
